feat: add app restart controls and centralize grant definitions
