+++ cluster/prov_opensvc_app.go
@@ -182,6 +182,11 @@ func (cluster *Cluster) OpenSVCStartAppService(app *App, node string) error {
 }
 
 func (cluster *Cluster) OpenSVCRestartAppService(app *App, node string, rid string) error {
+	if err := validateAppRestartRid(rid); err != nil {
+		cluster.LogModulePrintf(cluster.Conf.Verbose, config.ConstLogModOrchestrator, config.LvlErr, "App restart validation failed: %s", err)
+		return err
+	}
+
 	svc := cluster.OpenSVCConnect()
 	if svc.IsV3() {
 		if rid != "" {
